Fix #1629: improve next.js auth handling
