+++ src/frontend/next/src/interfaces/index.ts
@@ -1,10 +1,18 @@
-export type User = {
-  name?: string;
-  email?: string;
-  id?: string;
-  isAdmin?: boolean;
+export interface User {
   isLoggedIn: boolean;
-};
+}
+
+export interface AuthenticatedUser extends User {
+  name: string;
+  email: string;
+  id: string;
+  isAdmin: boolean;
+  isLoggedIn: true;
+}
+
+export interface AdminUser extends AuthenticatedUser {
+  isAdmin: true;
+}
 
 export type Feed = {
   id: string;
